#14 move custom modules to lib folder
