+++ lib/dbmanager.js
