+++ lib/takeOrder.js
