+++ lib/witlogic.js
